feat(adp): change counter folder to adp module and refactor counter.js
