+++ app/adp/plugins/counter/models.py
@@ -13,3 +13,4 @@ class Counter(CMSPlugin):
     measurement = models.CharField(verbose_name="Unidade de medida", max_length=140, null=True, blank=True)
     description = models.TextField(verbose_name="Descrição", null=True, blank=True)
     tooltip_text = models.TextField(verbose_name="Texto do Tooltip", null=True, blank=True)
+
