+++ app/adp/plugins/counter/static/counter/css/counter.scss
@@ -1,0 +1,3 @@
+.counter {
+  background-color: var(--bs-gray-500);
+}
