+++ app/adp/plugins/counter/templates/counter/plugins/counter.html
@@ -1,0 +1,38 @@
+{% load compress cms_tags static sekizai_tags %}
+
+{% addtoblock "css" %}
+  {% compress css %}
+  <style type="text/x-scss">
+    @import 'counter/css/counter.scss';
+  </style>
+  {% endcompress %}
+{% endaddtoblock %}
+
+{% addtoblock "js" %}
+  <script src="{% static "/counter/js/lib/countUp.min.js" %}" type="module"></script>
+  <script src="{% static "/counter/js/counter.js" %}" type="module"></script>
+{% endaddtoblock %}
+
+<div class="counter">
+  <div class="counter-number">
+    {% if instance.start_number or instance.end_number %}
+      <span class="counterNumber" data-start-num="{{ instance.start_number }}" data-end-num="{{ instance.end_number }}"></span>
+    {% endif %}
+
+    {% if instance.start_date or instance.end_date %}
+      <span class="counterDate" data-start-date="{{ instance.start_date|date:'Y-m-d' }}" data-end-date="{{ instance.end_date|date:'Y-m-d' }}"></span>
+    {% endif %}
+    {% if instance.measurement %}
+      <span>{{ instance.measurement }}</span>
+    {% endif %}
+  </div>
+    {% if instance.measurement %}
+    <p class="counter-description">{{ instance.description }}</p>
+    {% endif %}
+
+    {% if instance.tooltip_text %}
+      <div class="counter-tooltip">
+        <span class="counter-tooltip-text">{{ instance.tooltip_text }}</span>
+      </div>
+    {% endif %}
+</div>
