+++ app/adp/settings/base.py
@@ -64,10 +64,11 @@ INSTALLED_APPS = [
     "contrib.ds.blocks",
     "contrib.ds.card",
     "contrib.ds.carousel",
-    "contrib.ds.counter",
     "contrib.ds.grid",
     "contrib.ds.link",
     # Aplicativos por site
+    "adp.plugins.counter",
+
 ]
 
 # URLs
