--- app/contrib/ds/counter/templates/counter/plugins/counter.html
@@ -1,23 +1,0 @@
-{% load cms_tags static %}
-
-{% block js %}
-  <script src="{% static '/ds/js/lib/countUp.min.js' %}" type="module"></script>
-  <script src="{% static "/ds/js/counter.js" %}" type="module"></script>
-{% endblock %}
-
-<div class="counter">
-  <div class="counter-number">
-    {% if instance.start_number or instance.end_number %}
-      <span id="counterNumber" data-start-num="{{ instance.start_number }}" data-end-num="{{ instance.end_number }}"></span>
-    {% endif %}
-
-    {% if instance.start_date or instance.end_date %}
-      <span id="counterDate" data-start-date="{{ instance.start_date|date:'Y-m-d' }}" data-end-date="{{ instance.end_date|date:'Y-m-d' }}"></span>
-    {% endif %}
-      <span>{{ instance.measurement }}</span>
-  </div>
-    <p class="counter-description">{{ instance.description }}</p>
-    <div class="counter-tooltip">
-      <span class="counter-tooltip-text">{{ instance.tooltip_text }}</span>
-    </div>
-</div>
